Remaining fixups
