+++ include/ccf/service/signed_req.h
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #pragma once
+
+#include "ccf/crypto.h"
 #include "ds/json.h"
 
 #include <vector>
